split_video_by_second/requirements.txt
ffmpeg-python
numpy # Often a useful utility, and ffmpeg-python might have it as a transitive dep or it's good for math. 